requirements.txt
streamlit
easyocr
transformers
torch
torchvision
opencv-python-headless
Pillow
streamlit-image-coordinates
huggingface-hub